Added new methods to the existing native bridge functionality in the main sdk and more
Added a delay option to regulate the time between requests to the blockchain for transaction confirmation.
Added backward compatibility to the sdk, for a smooth transition between old and new versions.
Added extension methods for the 'decimal' data type, which can be used to call `ToNanoton()` and `FromNanoton()` methods at once.

[Demo] Updated the scene with an example of using sdk for the current version of the functionality.
[Bug fix] Fixed a critical bug when generating application config at build time, when a necessary slash in a link could be missing, which could prevent the image or the whole config from loading.

+++ Uniton Connect/Runtime/Data/Message/InMessageData.cs
@@ -23,10 +23,10 @@ namespace UnitonConnect.Core.Data
         public bool IsBounced { get; set; }
 
         [JsonProperty("value")]
-        public long Value { get; set; }
+        public decimal Value { get; set; }
 
         [JsonProperty("fwd_fee")]
-        public long ForwardFee { get; set; }
+        public decimal ForwardFee { get; set; }
 
         [JsonProperty("ihr_fee")]
         public long IhrFee { get; set; }
